Add initial data classes
The program will work with computing classes passing each other an instance
of these.

Data class Input holds the raw data yet to validate into an ip and mask
and is what the validator expects to work with.
Data class ValidatedData holds the data verified by the Validator and is
what the Calculator will work with to produce a Results instance.
Data class Results holds the product of all our work and is what Main
will end up using to print the results back to the command line.