Add the Main class
Entry point for the program, first to be added so I can test the other
classes as I add new features to them. Right now it only prints some
hardcoded data to test that Results and the displayResults() method work
as intended. All that will dissapear as the program progresses.