Add the method populateFields() to the Input class
The Input class receives the raw input of our program, and uses the
populateFields to try and fill its own fields, being naive and assuming
that the raw input will have the ip and mask separated with a '/'. It
has to be validated still, but it hopes for the best, and so long as
the user separated the ip and the mask in the proper way, this should
work flawlessly without much problems.