Modify the shell script so it uses functions
Now the script can be sourced and the function run can be run writing
'run <args>' in the prompt, as opposed to writing './run.sh <args>' as
before. This is the first time I'm actually making use of this, despite
knowing about it for forever, and now I love it. How could I live
without it before is a mystery